requirements.txt
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
sentence-transformers
streamlit
docx2txt
python-pptx
Pillow
llama-parse
llama-index-readers-file
python-dotenv